+++ src/notifiers/mod.rs
@@ -3,6 +3,9 @@ use crate::reactor::State;
 mod telegram;
 pub use self::telegram::TelegramNotifier;
 
+mod command;
+pub use self::command::CommandNotifier;
+
 #[derive(Debug)]
 pub struct Notification {
     pub checker_id: String,